// src/app/models/paciente.ts
export interface Paciente {
  fecha_reporte_web?: string;
  id_de_caso?: string;
  fecha_de_notificaci_n?: string;
  departamento?: string;
  departamento_nom?: string;
  ciudad_municipio?: string;
  ciudad_municipio_nom?: string;
  edad?: string;
  unidad_medida?: string;
  sexo?: string;
  fuente_tipo_contagio?: string;
  ubicacion?: string;
  estado?: string;
  pais_viajo_1_cod?: string;
  pais_viajo_1_nom?: string;
  recuperado?: string;
  fecha_inicio_sintomas?: string;
  fecha_muerte?: string;
  fecha_diagnostico?: string;
  fecha_recuperado?: string;
  tipo_recuperacion?: string;
  per_etn_?: string;
  nom_grupo_?: string;
}